Refactor: Split components. Use tmi.js for chat detection.
