+++ src/components/Emotes/Toy.js
@@ -1,363 +1,35 @@
-/* eslint-disable no-bitwise */
-/* eslint-disable no-param-reassign */
-
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-import p2 from 'p2';
+import React from 'react';
+import { Stage } from '@inlet/react-pixi';
 import * as PIXI from 'pixi.js';
-import moment from 'moment';
+
 import styled from 'styled-components';
 
 import { useMessageContext, useNotificationContext } from 'providers';
 
-import idMap from './assets/idMap.json';
-import sprites from './hype.json';
+import Scene from './Scene';
 
 PIXI.utils.skipHello();
-
-const AUTO_SPAWN_FOR_TEST = false;
-const { PUBLIC_URL } = process.env;
-
-const SCALE = 0.5;
-const IMG_WIDTH = 112 * SCALE;
-const MAX_OBJECTS = 200;
-const MAX_AGE = 180;
-const HYPE_DURATION = 30;
-
-const OBJECT = 2 ** 0;
-const WALL = 2 ** 1;
-const GROUND = 2 ** 2;
-
 const WIDTH = 1920;
 const HEIGHT = 1080;
 
 function Toy() {
-  const [notifications] = useNotificationContext();
-  const [messages] = useMessageContext();
-  const [hypeActive, setHypeActive] = useState(false);
-  const [spamActive, setSpamActive] = useState(true);
-
-  // Instance Variables.
-  const hypeTimer = useRef(null);
-  const lastTime = useRef(null);
-  const objectList = useRef([]);
-  const stageRef = useRef(null);
-  const textures = useRef([]);
-
-  // Construction.
-  const world = new p2.World({ gravity: [0, -9.82] });
-  const loader = new PIXI.Loader();
-  const pixiApp = new PIXI.Application({
-    width: WIDTH,
-    height: HEIGHT,
-    transparent: true,
-    resolution: window.devicePixelRatio || 1
-  });
-  const pixiContainer = new PIXI.ParticleContainer(3000, {
-    scale: true,
-    rotation: true
-  });
-
-  const shuffleArray = (array, num) => {
-    let target = num;
-    if (num > array.length) {
-      target = array.length;
-    }
-    const random = array.sort(() => 0.5 - Math.random()).slice(0, target);
-    return random;
-  };
-
-  const createObject = useCallback(
-    (auto, cheer, emote) => {
-      const object = {
-        alive: true,
-        created: moment(),
-        render: null,
-        physics: null
-      };
-
-      const x = Math.random() * WIDTH;
-      const y = -1 * (IMG_WIDTH / 2 + 10);
-
-      const angle = Math.random() * 360 * (Math.PI / 180);
-      const angularVelocity = Math.random() * 5;
-
-      const circleBody = new p2.Body({
-        mass: Math.random() * 95 + 5,
-        position: [x, y],
-        angle,
-        angularVelocity
-      });
-      const circleShape = new p2.Circle({ radius: IMG_WIDTH / 2 });
-      circleBody.addShape(circleShape);
-      circleShape.collisionGroup = OBJECT;
-      circleShape.collisionMask = OBJECT | WALL | GROUND;
-
-      object.physics = circleBody;
-      world.addBody(circleBody);
-
-      let spriteImg;
-
-      if (emote) {
-        spriteImg = textures.current[`${emote}.png`];
-      } else {
-        const emoteName =
-          sprites[Math.floor(Math.random() * (sprites.length - 1))];
-        spriteImg = textures.current[`${emoteName}.png`];
-      }
-
-      const blankObj = new PIXI.Sprite(spriteImg);
-      blankObj.anchor.set(0.5);
-      blankObj.scale.x = SCALE;
-      blankObj.scale.y = SCALE;
-      blankObj.x = x;
-      blankObj.y = y;
-      blankObj.rotation = angle;
-
-      object.render = blankObj;
-
-      pixiContainer.addChild(blankObj);
-
-      objectList.current.push(object);
-
-      if (auto && AUTO_SPAWN_FOR_TEST) {
-        setTimeout(() => createObject(true, false), Math.random() * 2 * 1000);
-      }
-    },
-    [pixiContainer, world]
+  return (
+    <Wrapper>
+      <Stage width={WIDTH} height={HEIGHT} options={{ transparent: true }}>
+        <useMessageContext.Provider>
+          <useNotificationContext.Provider>
+            <Scene width={WIDTH} height={HEIGHT} />
+          </useNotificationContext.Provider>
+        </useMessageContext.Provider>
+      </Stage>
+    </Wrapper>
   );
-
-  const handleSupport = useCallback(
-    inputValue => {
-      const value = parseInt(inputValue, 10);
-      if (value >= 500) {
-        setHypeActive(true);
-        clearTimeout(hypeTimer.current);
-        hypeTimer.current = setTimeout(
-          setHypeActive(false),
-          HYPE_DURATION * 1000
-        );
-      }
-      const timesIntermediary = Math.round(Math.max(1, value / 100));
-      const times = timesIntermediary;
-      for (let i = 0; i < times; i += 1) {
-        setTimeout(() => createObject(false), (i + 1) * 250);
-      }
-    },
-    [createObject]
-  );
-
-  const handleEmotes = useCallback(
-    emotes => {
-      if (spamActive || hypeActive) {
-        const acceptedEmotes = [];
-        Array.from(emotes.keys()).forEach(emote => {
-          if (idMap[emote]) {
-            for (let i = 0; i < emotes.get(emote).length; i += 1) {
-              acceptedEmotes.push(idMap[emote]);
-            }
-          }
-        });
-        shuffleArray(acceptedEmotes, 15).forEach((emote, id) => {
-          setTimeout(() => createObject(false, false, emote), (id + 1) * 350);
-        });
-      }
-    },
-    [spamActive, hypeActive, createObject]
-  );
-
-  const calculateSubValue = plan => {
-    let value = 100;
-    switch (plan) {
-      case 'Prime':
-      case '1000':
-        value *= 5;
-        break;
-      case '2000':
-        value *= 10;
-        break;
-      case '3000':
-        value *= 25;
-        break;
-      default:
-        value = 0;
-        break;
-    }
-
-    return value;
-  };
-
-  const draw = () => {
-    objectList.current.forEach(object => {
-      [object.render.x] = object.physics.interpolatedPosition;
-      object.render.y = -1 * object.physics.interpolatedPosition[1];
-      object.render.rotation = object.physics.interpolatedAngle;
-    });
-  };
-
-  const animate = useCallback(
-    time => {
-      const { current } = lastTime;
-
-      requestAnimationFrame(animate);
-      const deltaTime = current ? (time - current) / 1000 : 0;
-      world.step(1 / 60, deltaTime, 10);
-      draw();
-      lastTime.current = time;
-    },
-    [world]
-  );
-
-  const cleanup = useCallback(() => {
-    const { current } = objectList;
-    const outOfBounds = -1 * (HEIGHT + IMG_WIDTH / 2 + 10);
-
-    if (current.length > MAX_OBJECTS) {
-      current.forEach((object, id) => {
-        if (object.alive) {
-          const age = moment().diff(object.created, 's');
-          if (age > MAX_AGE / (4 * (current.length / MAX_OBJECTS))) {
-            object.alive = false;
-            object.physics.shapes[0].collisionMask = OBJECT | WALL;
-          }
-        } else if (object.physics.position[1] <= outOfBounds) {
-          pixiApp.stage.removeChild(object.render);
-          object.render.destroy();
-          world.removeBody(object.physics);
-          current.splice(id, 1);
-        }
-      });
-    } else {
-      current.forEach((object, id) => {
-        if (object.alive) {
-          const age = moment().diff(object.created, 's');
-          if (age > MAX_AGE) {
-            object.alive = false;
-            object.physics.shapes[0].collisionMask = OBJECT | WALL;
-          }
-        } else if (object.physics.position[1] <= outOfBounds) {
-          pixiApp.stage.removeChild(object.render);
-          object.render.destroy();
-          world.removeBody(object.physics);
-          current.splice(id, 1);
-        }
-      });
-    }
-  }, [pixiApp, world]);
-
-  const buildPhysics = useCallback(() => {
-    world.setGlobalStiffness(1e4);
-    world.setGlobalRelaxation(2);
-
-    const groundBody = new p2.Body({
-      mass: 0,
-      position: [WIDTH / 2, -1 * (HEIGHT + 10)]
-    });
-    const groundShape = new p2.Box({
-      width: WIDTH,
-      height: 10
-    });
-    groundBody.addShape(groundShape);
-    world.addBody(groundBody);
-
-    groundShape.collisionGroup = GROUND;
-
-    const leftBody = new p2.Body({
-      mass: 0,
-      position: [-10, -1 * (HEIGHT / 2)]
-    });
-    const leftShape = new p2.Box({
-      width: 10,
-      height: HEIGHT
-    });
-    leftBody.addShape(leftShape);
-    world.addBody(leftBody);
-
-    leftShape.collisionGroup = WALL;
-
-    const rightBody = new p2.Body({
-      mass: 0,
-      position: [WIDTH + 10, -1 * (HEIGHT / 2)]
-    });
-    const rightShape = new p2.Box({
-      width: 10,
-      height: HEIGHT
-    });
-    rightBody.addShape(rightShape);
-    world.addBody(rightBody);
-
-    rightShape.collisionGroup = WALL;
-
-    const topBody = new p2.Body({
-      mass: 0,
-      position: [WIDTH / 2, -10]
-    });
-    const topShape = new p2.Box({
-      width: WIDTH,
-      height: 10
-    });
-    topBody.addShape(topShape);
-    world.addBody(topBody);
-
-    topShape.collisionGroup = WALL;
-
-    world.on('postStep', () => cleanup());
-  }, [cleanup, world]);
-
-  const buildStage = useCallback(() => {
-    stageRef.current.appendChild(pixiApp.view);
-    pixiContainer.width = WIDTH;
-    pixiContainer.height = HEIGHT;
-    pixiApp.stage.addChild(pixiContainer);
-  }, [stageRef, pixiApp, pixiContainer]);
-
-  useEffect(() => {
-    const emoteTexture = `${PUBLIC_URL}/sprites/emoteTexture.json`;
-    loader.add(emoteTexture).load((_loader, resources) => {
-      textures.current = resources[emoteTexture].textures;
-      buildPhysics();
-      buildStage();
-      requestAnimationFrame(animate);
-    });
-
-    return () => {
-      clearTimeout(hypeTimer.current);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (AUTO_SPAWN_FOR_TEST) {
-      setTimeout(() => createObject(true, false), 3 * 1000);
-    }
-  }, [createObject]);
-
-  useEffect(() => {
-    const [message] = messages.slice(-1);
-    if (message) {
-      const { emoteOffsets } = message.tags;
-      if (emoteOffsets) handleEmotes(emoteOffsets);
-    }
-  }, [messages, handleEmotes]);
-
-  useEffect(() => {
-    const [notification] = notifications;
-    if (notification && notification.bucket === 'subscription') {
-      handleSupport(calculateSubValue(notification.plan));
-    }
-  }, [notifications, handleSupport]);
-
-  useEffect(() => {
-    if (AUTO_SPAWN_FOR_TEST) {
-      setTimeout(() => createObject(true, false), 3 * 1000);
-    }
-  }, [createObject]);
-
-  return <Wrapper ref={stageRef} />;
 }
 
 const Wrapper = styled.div`
   position: absolute;
-  width: 100vw;
-  height: 100vh;
+  width: 1920px;
+  height: 1080px;
 `;
 
 export default Toy;
